TMF622/source/ProductOrderCaptureValidation/requirements.txt
fastapi>=0.110,<1
uvicorn>=0.27,<1
httpx>=0.27,<1
pymongo>=4.6,<5
prometheus-client>=0.20,<1
python-dotenv>=1.0,<2
mcp>=1.0,<2
starlette>=0.37,<1

